+++ src/data/tests/data/processed/maps/osm_signals.cpg
@@ -1,0 +1,1 @@
+ISO-8859-1
